consolidate util files
